Unify function to slugify titles
We have multiple utility functions to slugify or "normalize" the
channel title or content title. Here is a first attempt to unify them.

Their results were different. So also rename two overrides and one
asset.

https://phabricator.endlessm.com/T32348
