+++ packages/template-ui/channel-overrides/sikana-english/assets/sikana.png
