+++ packages/template-ui/src/utils.js
@@ -1,17 +1,5 @@
 import arrayToTree from 'array-to-tree';
 
-export function getSlug(title) {
-  return title
-    .toLowerCase()
-    .replace(/&/g, 'and')
-    // Remove invalid characters
-    .replace(/[^a-z0-9 -]/g, '')
-    // Replace whitespace by -
-    .replace(/\s+/g, '-')
-    // Collapse dashes
-    .replace(/-+/g, '-');
-}
-
 export function getNodesTree(nodes) {
   return arrayToTree(nodes, { parentProperty: 'parent' });
 }
